Inference

+++ Fuzzy Logic/src/com/genetics/Point.java
@@ -1,10 +1,10 @@
 package com.genetics;
 
 public class Point {
-    public  int x;
-    public int y;
+    public  float x;
+    public float y;
 
-    public Point(int x,int y){
+    public Point(float x,float y){
         this.x=x;
         this.y=y;
     }
