+++ Fuzzy Logic/src/com/genetics/Shape.java
@@ -6,12 +6,12 @@ public class Shape {
     Point[] points;
 
 
-    public float fuzzification(int value){
+    public float fuzzification(float value){
         for(int i=0;i<points.length-1;i++){
             if(points[i].x!=points[i+1].x) {
-                float slope = ((float) points[i + 1].y - (float) points[i].y) / ((float) points[i + 1].x - (float) points[i].x);
-                float b = (float) points[i].y - (slope * (float) points[i].x);
-                float y = slope * (float) value + b;
+                float slope = (points[i + 1].y - points[i].y) / (points[i + 1].x - points[i].x);
+                float b = points[i].y - (slope * points[i].x);
+                float y = slope * value + b;
                 if (y <= 1 && y > 0)
                     return y;
             }
