+++ Fuzzy Logic/src/com/genetics/Triangular.java
@@ -2,7 +2,7 @@ package com.genetics;
 
 public class Triangular extends Shape {
 
-    public Triangular(String name,int[] points){
+    public Triangular(String name,float[] points){
         this.name=name;
         this.points=new Point[3];
         for(int i=0;i<points.length;i++){
